LED_Display/requirements.txt
Pillow==9.2.0
python-dotenv==0.21.0
